pd: Major refactor to internal structure
This change refactors the internal structure of `pd` by breaking up the
monolithic ABCI service into four component services, one for each ABCI
category:

* a `Consensus` server processing `ConsensusRequest`s;
* a `Mempool` server processing `MempoolRequest`s (`CheckTx`);
* an `Info` server processing `InfoRequest`s;
* a (currently unused) `Snapshot` server for snapshot handling (which we don't implement).

In addition, we break up the previous `State` structure into `state::Reader`
and `state::Writer` structures.  This means that writes can be more logically
isolated to the `state::Writer::{commit_genesis, commit_block}` methods, and
that the consensus-critical writer role can use a distinct connection pool from
the non-critical reader role.

It also allows us to supplement the database (the primary means of shared
state) with `tokio::sync::watch` channels (the single-producer sender end is
held by the `state::Writer`, and the multi-consumer receiver end is held by the
`state::Reader`s).  This allows synchronized in-memory caching and allows
receivers to be notified on state changes.  For instance, the `Mempool` server
now holds a `watch` channel for the block height, so it can reset the mempool
nullifier set on each block commit in anticipation of re-checks; the wallet
server could use the watch channel for the `ChainParams` to implement chain-id
consistency checks, etc.

The biggest part of (and original motivation for) this change is the
`Consensus` server.  Previously, we handled consensus messages as part of the
monolithic `App`'s `Service` implementation.  Because our state access is
asynchronous, this meant that our actual logic had to be in the returned
futures, which led to messy code to (a) share `App` state with those futures
and then (b) re-impose sequential execution on some (but not all) of them.

Instead, the new code creates a dedicated `consensus::Worker` task; the
`Consensus` `Service` implementation sends requests to the worker task, which
processes them sequentially and statefully, dramatically simplifying that code
(e.g., the `Worker` can simply hold a `PendingBlock` directly and edit it,
rather than having to do an `Arc<Mutex<...>>` dance).  This change was already
big enough, but a logical next step in the evolution of the `Worker` code would
be to try to disentangle its state update handling code into functional areas
(e.g., staking, shielded pool, etc).
