+++ pd/Cargo.toml
@@ -32,6 +32,7 @@ ics23 = { git = "https://github.com/penumbra-zone/ics23" }
 
 
 # External dependencies
+evmap = "10"
 async-stream = "0.2"
 bincode = "1.3.3"
 blake2b_simd = "0.5"
@@ -39,7 +40,7 @@ bytes = "1"
 chrono = { version = "0.4", default-features = false, features = ["serde"] }
 csv = "1.1"
 directories = "4.0"
-tokio = { version = "1", features = ["full"]}
+tokio = { version = "1.16", features = ["full"]}
 tokio-stream = "0.1"
 tokio-util = "0.6"
 tower = { version = "0.4", features = ["full"]}
