+++ pd/src/lib.rs
@@ -4,23 +4,28 @@
 #![recursion_limit = "512"]
 #![allow(clippy::clone_on_copy)]
 
-mod app;
+mod consensus;
 mod db;
+mod info;
+mod mempool;
 mod pd_metrics;
 mod pending_block;
 mod request_ext;
-mod sequential;
-mod state;
+mod snapshot;
 mod verify;
 mod wallet;
 
-use sequential::Sequencer;
-
 pub mod genesis;
+pub mod state;
 pub mod testnet;
 
-pub use app::App;
+pub use consensus::Consensus;
+pub use info::Info;
+pub use mempool::Mempool;
 pub use pd_metrics::register_all_metrics;
-pub use pending_block::PendingBlock;
-pub use request_ext::RequestExt;
-pub use state::State;
+use pending_block::PendingBlock;
+use request_ext::RequestExt;
+pub use snapshot::Snapshot;
+
+/// The age limit, in blocks, on anchors accepted in transaction verification.
+pub const NUM_RECENT_ANCHORS: usize = 256;
